requirements.txt
numpy
scipy
torch==0.4.1.post2
torchvision
scikit-learn
pyyaml
tqdm
ipython
